Refactor to allow for a distinct provider conformance test
This commit renames the 'conformance' package to 'crossplane', to indicate that
it will be the package that tests Crossplane conformance. A future commit will
add a distinct 'provider' conformance test package.
