+++ go.mod
@@ -4,7 +4,6 @@ go 1.16
 
 require (
 	github.com/crossplane/crossplane v1.2.0
-	github.com/crossplane/crossplane-runtime v0.13.0
 	github.com/pkg/errors v0.9.1
 	k8s.io/api v0.20.2
 	k8s.io/apiextensions-apiserver v0.20.1
